Calculator- July'17 Long Challenge
Chef has a calculator which has two screens and two buttons. Initially,
each screen shows the number zero. Pressing the first button increments
the number on the first screen by 1, and each click of the first button
consumes 1 unit of energy.

Pressing the second button increases the number on the second screen by
the number which is currently appearing on the first screen. Each click
of the second button consumes B units of energy.

Initially the calculator has N units of energy.

Now chef wonders what the maximum possible number is, that he gets on
the second screen of the calculator, with the limited energy.

Input
The first line of the input contains an integer T denoting the number of
test cases.

Each test case is described using a single line containing two integers,
N and B.

Output
For each test case, output a single line containing the answer to this
test case.

Constraints
1 ≤ T ≤ 10,000
1 ≤ N, B ≤ 1,000,000,000